[docs] API reference spacing reduce and display improvements
